Professional Email: Hide optional fields in the "Add mailbox" form for consistency with the Email Comparison view (#66106)
* Export password reset tip field as a standalone component

* Hide redundant fields in email form to reduce user journey

* Add default value for password reset field

* Import variables css
