+++ client/my-sites/email/add-mailboxes/index.tsx
@@ -2,7 +2,7 @@ import { Card } from '@automattic/components';
 import { useShoppingCart } from '@automattic/shopping-cart';
 import { TranslateResult, useTranslate } from 'i18n-calypso';
 import page from 'page';
-import { PropsWithChildren, useState } from 'react';
+import { MouseEvent, PropsWithChildren, useState } from 'react';
 import { useSelector } from 'react-redux';
 import DocumentHead from 'calypso/components/data/document-head';
 import QueryProductsList from 'calypso/components/data/query-products-list';
@@ -29,10 +29,15 @@ import {
 import TitanUnusedMailboxesNotice from 'calypso/my-sites/email/add-mailboxes/titan-unused-mailboxes-notice';
 import EmailHeader from 'calypso/my-sites/email/email-header';
 import { NewMailBoxList } from 'calypso/my-sites/email/form/mailboxes/components/new-mailbox-list';
+import PasswordResetTipField from 'calypso/my-sites/email/form/mailboxes/components/password-reset-tip-field';
 import getMailProductForProvider from 'calypso/my-sites/email/form/mailboxes/components/selectors/get-mail-product-for-provider';
 import getCartItems from 'calypso/my-sites/email/form/mailboxes/components/utilities/get-cart-items';
 import { getEmailProductProperties } from 'calypso/my-sites/email/form/mailboxes/components/utilities/get-email-product-properties';
 import { MailboxOperations } from 'calypso/my-sites/email/form/mailboxes/components/utilities/mailbox-operations';
+import {
+	FIELD_ALTERNATIVE_EMAIL,
+	FIELD_NAME,
+} from 'calypso/my-sites/email/form/mailboxes/constants';
 import { EmailProvider } from 'calypso/my-sites/email/form/mailboxes/types';
 import { INBOX_SOURCE } from 'calypso/my-sites/email/inbox/constants';
 import {
@@ -40,6 +45,7 @@ import {
 	emailManagementInbox,
 	emailManagementTitanSetUpMailbox,
 } from 'calypso/my-sites/email/paths';
+import { getCurrentUserEmail } from 'calypso/state/current-user/selectors';
 import { ProductListItem } from 'calypso/state/products-list/selectors/get-products-list';
 import getCurrentRoute from 'calypso/state/selectors/get-current-route';
 import {
@@ -48,6 +54,7 @@ import {
 	isRequestingSiteDomains,
 } from 'calypso/state/sites/domains/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
+import type { HiddenFieldNames } from 'calypso/my-sites/email/form/mailboxes/components/new-mailbox-list';
 import type { SiteData } from 'calypso/state/ui/selectors/site-data';
 import type { translate } from 'i18n-calypso';
 
@@ -221,8 +228,13 @@ const MailboxesForm = ( {
 	emailProduct: ProductListItem | null;
 	goToEmail: () => void;
 } ): JSX.Element => {
+	const userEmail = useSelector( getCurrentUserEmail );
 	const [ isAddingToCart, setIsAddingToCart ] = useState( false );
 	const [ isValidating, setIsValidating ] = useState( false );
+	const [ hiddenFieldNames, setHiddenFieldNames ] = useState< HiddenFieldNames[] >( [
+		FIELD_NAME,
+		FIELD_ALTERNATIVE_EMAIL,
+	] );
 
 	const cartKey = useCartKey();
 	const cartManager = useShoppingCart( cartKey );
@@ -230,6 +242,11 @@ const MailboxesForm = ( {
 	if ( isLoadingDomains || ! emailProduct ) {
 		return <AddEmailAddressesCardPlaceholder />;
 	}
+
+	const showAlternateEmailField = ( event: MouseEvent< HTMLElement > ) => {
+		event.preventDefault();
+		setHiddenFieldNames( [ FIELD_NAME ] );
+	};
 
 	const onCancel = () => {
 		recordClickEvent( {
@@ -290,6 +307,8 @@ const MailboxesForm = ( {
 			<Card>
 				<NewMailBoxList
 					areButtonsBusy={ isAddingToCart || isValidating }
+					hiddenFieldNames={ hiddenFieldNames }
+					initialFieldValues={ { [ FIELD_ALTERNATIVE_EMAIL ]: userEmail } }
 					onSubmit={ onSubmit }
 					onCancel={ onCancel }
 					provider={ provider }
@@ -297,7 +316,11 @@ const MailboxesForm = ( {
 					showAddNewMailboxButton
 					showCancelButton
 					submitActionText={ translate( 'Continue' ) }
-				/>
+				>
+					{ hiddenFieldNames.includes( FIELD_ALTERNATIVE_EMAIL ) && (
+						<PasswordResetTipField tipClickHandler={ showAlternateEmailField } />
+					) }
+				</NewMailBoxList>
 			</Card>
 		</>
 	);
