+++ client/my-sites/email/email-providers-comparison/stacked/provider-cards/professional-email-card.scss
@@ -12,7 +12,3 @@
 		padding-right: 12px;
 	}
 }
-
-.professional-email-card__change-it-button {
-	color: var( --color-link );
-}
