+++ client/my-sites/email/form/mailboxes/components/password-reset-tip-field/style.scss
@@ -1,0 +1,5 @@
+@import '@wordpress/base-styles/variables';
+
+.password-reset-tip-field__change-it-button {
+	color: var( --color-link );
+}
